Agrego logica de owner

+++ src/controllers/products.controller.js
@@ -5,14 +5,14 @@ const productsService = new ProductsService();
 export const getProducts = async (req, res) => {
     try {
 
-        const products = await productsService.getProducts(req.query);        
+        const products = await productsService.getProducts(req.query);
         res.status(200).json(products);
 
     } catch (error) {
         console.error(error);
-        res.status(500).json({error: "Error loading the products. "+error.message});
+        res.status(500).json({ error: "Error loading the products. " + error.message });
     }
-    
+
 }
 
 export const getProductById = async (req, res) => {
@@ -24,14 +24,17 @@ export const getProductById = async (req, res) => {
 
     } catch (error) {
         console.error(error);
-        res.status(400).json({ error: "Error searching the product. "+error.message });
+        res.status(400).json({ error: "Error searching the product. " + error.message });
     }
-    
+
 }
 
 export const createProduct = async (req, res, next) => {
     try {
         let newProduct = req.body;
+        if (req.user.role === "premium") {
+            newProduct.owner = req.user.email;
+        }
         let productCreated = await productsService.createProduct(newProduct);
 
         res.status(200).json(productCreated);
@@ -39,7 +42,7 @@ export const createProduct = async (req, res, next) => {
     } catch (error) {
         next(error);
     }
-    
+
 }
 
 export const updateProduct = async (req, res) => {
@@ -48,15 +51,25 @@ export const updateProduct = async (req, res) => {
         const productId = req.params.pid;
         const productFields = req.body;
 
+        const existProduct = await productsService.getProductById(productId)
+        if (!existProduct) {
+            return res.status(401).json({ status: "error", message: "The product doesn't exist" });
+        }
+
+        if (req.user.role === "premium") {
+            if (req.user.email != existProduct.owner)
+            return res.status(403).json({ status: "error", message: "Product Owner or Admin role required" });
+        }
+
         let productUpdated = await productsService.updateProduct(productId, productFields);
 
         res.status(200).json(productUpdated);
-        
+
     } catch (error) {
         console.error(error);
-        res.status(400).json({ error: "Error updating product. "+error.message });
+        res.status(400).json({ error: "Error updating product. " + error.message });
     }
-    
+
 }
 
 
@@ -65,12 +78,22 @@ export const deleteProduct = async (req, res) => {
 
         const productId = req.params.pid;
 
+        const existProduct = await productsService.getProductById(productId)
+        if (!existProduct) {
+            return res.status(401).json({ status: "error", message: "The product doesn't exist" });
+        }
+
+        if (req.user.role === "premium") {
+            if (req.user.email != existProduct.owner)
+            return res.status(403).json({ status: "error", message: "Product Owner or Admin role required" });
+        }
+
         let productDeleted = await productsService.deleteProduct(productId);
         res.status(200).json(productDeleted);
 
     } catch (error) {
         console.error(error);
-        res.status(400).json({ error: "Error deleting product. "+error.message });
+        res.status(400).json({ error: "Error deleting product. " + error.message });
     }
-    
+
 }
